refactor: 정적파일타입 enum 패키지 이동
