+++ src/test/java/webserver/filter/FileExtensionTypeTest.java
@@ -1,4 +1,4 @@
-package webserver.requestmapping;
+package webserver.filter;
 
 import static org.assertj.core.api.Assertions.*;
